Code for Release 02 (Temperature Sensor and Error Handling).

+++ src/OLED/oled_module.h
@@ -59,7 +59,7 @@ Date		  	Version		Author		Short Task Description (specify task ID if available)
 /**                     EXPORTED FUNCTIONS                                 **/
 /**                                                                        **/
 /****************************************************************************/
-void Oled_intToString(int value, uint8_t* pBuf, uint32_t len, uint32_t base);
+void Oled_intToString(int_fast8_t value, uint8_t* pBuf, uint32_t len, uint32_t base);
 
 #endif /* _OLED_MODULE_INCLUDED */
